+++ core/SolverStrategy.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
-
+from PySide6.QtWidgets import QWidget
 
 class SolverStrategy:
     @abstractmethod
-    def solve(self, objectiveFunction, constraints):
+    def solve(self, objectiveFunction, constraints)-> QWidget:
         pass