+++ DocFIles/CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog 📃
 
+## V1.1.0 - 2026-05-11
+### Added
+- New result view, visualice the last table iteration, the objective function value and the non-zero varaibles values
+### Fixed
+- Repeated rows building the tableau
+
+### Refactored
+- Added a QStackedWidget for working differents views in a single MainWindow
+
 ## V1.0.0 - 2026-05-06
 First version of the software